base/serializers/post_serializers.py
from rest_framework import serializers


from base.models.post_models import Post

class PostSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Post
        fields = ['title', 'text', 'source', 'video']