requirements.txt
django==3.2.5
djangorestframework
django-environ==0.4.5
Pillow==9.1.0
libmagic==1.0
python-magic==0.4.27
psycopg2==2.9.5
gunicorn==20.1.0